pyproject.toml
[project]
name = "uv-redis-django-docker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.11",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
